refactor(dashboard): updated dashboard, login, & trip controllers to save user_id & to use it to retrieve trip data
