+++ controllers/api/trips-controller.js
@@ -5,9 +5,9 @@
 // =============================================================
 const router = require('express').Router();
 const multer = require('multer');
-const { Trip, TripSection, ItineraryItem } = require('../models');
-const loginAuth = require('../utils/auth');
-const userIdAuth = require('../utils/userIdAuth');
+const { Trip, TripSection, ItineraryItem } = require('../../models');
+const loginAuth = require('../../utils/auth');
+const userIdAuth = require('../../utils/userIdAuth');
 // =============================================================
 
 // Setting up folder to receive and format uploads via multer
