+++ controllers/home-controller.js
@@ -14,6 +14,7 @@ router.get('/', async (req, res) => {
     try {
         res.render('homepage', {
             loggedIn: req.session.loggedIn,
+            user_id: req.session.user_id,
         });
     } catch (err) {
         res.status(500).json(err);
@@ -24,6 +25,8 @@ router.get('/', async (req, res) => {
 // Get route for the dashboard page
 // =============================================================
 router.get('/trips', loginAuth, async (req, res) => {
+    console.log('getting trips data');
+    console.log(req.session.user_id);
     try {
         const tripData = await Trip.findAll({
             where: {
@@ -32,9 +35,11 @@ router.get('/trips', loginAuth, async (req, res) => {
             order: [['end_date', 'DESC']],
         });
         const trips = tripData.map((trip) => trip.get({ plain: true }));
+        console.log(trips);
         res.render('dashboard', {
             trips,
             loggedIn: req.session.loggedIn,
+            user_id: req.session.user_id,
         });
     } catch (err) {
         res.status(500).json(err);
