+++ controllers/trip-controller.js
@@ -123,6 +123,7 @@ router.get('/view/:id', [loginAuth, userIdAuth], async (req, res) => {
             sections,
             expenses,
             loggedIn: req.session.loggedIn,
+            user_id: req.session.user_id,
         });
     } catch (error) {
         console.error(error);
@@ -133,7 +134,7 @@ router.get('/view/:id', [loginAuth, userIdAuth], async (req, res) => {
 
 // Get edit page for a trip
 // =============================================================
-router.get('/edit/:id', loginAuth, async (req, res) => {
+router.get('/edit/:id', [loginAuth, userIdAuth], async (req, res) => {
     try {
         // Retrieve the trip with the specified ID and include its sections and items
         const trip = await Trip.findByPk(req.params.id, {
@@ -217,6 +218,7 @@ router.get('/edit/:id', loginAuth, async (req, res) => {
             image,
             sections,
             loggedIn: req.session.loggedIn,
+            user_id: req.session.user_id,
         });
     } catch (error) {
         console.error(error);
@@ -267,6 +269,7 @@ router.get('/create-trip', loginAuth, async (req, res) => {
             image,
             sections,
             loggedIn: req.session.loggedIn,
+            user_id: req.session.user_id,
         });
     } catch (error) {
         console.error(error);
