+++ views/partials/header.handlebars
@@ -13,7 +13,7 @@
                 <a href='#' class='item' id='scroll'>About Us</a>
             {{/unless}}
             {{#if loggedIn}}
-                <a href='/dashboard' class='item'>Dashboard</a>
+                <a href='/trips' class='item'>Dashboard</a>
             {{/if}}
         </div>
